backend/requirements.txt
Django==3.1.4
Pillow==8.0.1
asgiref==3.3.1
django-cors-headers==3.6.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
pip==20.3.3
pytz==2020.4
setuptools==51.0.0
sqlparse==0.4.1